Add files via upload
Reverse Polish notation (RPN) is a mathematical notation in which every operator follows all of its operands, for example "3 + 4 =" would be presented as "3 4 +". This code implements a calculator that performs addition, shift left, shift right, copying and printing. As part of the assignment I implemented a stack to hold input from the user where each number is represented in a linked list.